# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ShopXperience API",
    "description": "Multi-tenant Point of Sale System API",
    "version": "1.0.0",
    "contact": {
      "name": "ShopXperience Support",
      "email": "support@shopxperience.com"
    }
  },
  "servers": [
    {
      "url": "https://api.shopxperience.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Customer": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "first_name": { "type": "string" },
          "last_name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "phone": { "type": "string" },
          "address": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "Location": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "address": { "type": "string" },
          "phone": { "type": "string" },
          "tax_rate": { "type": "number", "format": "decimal" },
          "pricing_multiplier": { "type": "number", "format": "decimal" },
          "timezone": { "type": "string" },
          "currency": { "type": "string" },
          "is_active": { "type": "boolean" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    }
  },
  "paths": {
...
    "/api/customers": {
      "get": {
        "summary": "Get all customers",
        "description": "Retrieve paginated list of customers with optional search",
        "tags": ["Customers"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Customers retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "customers": { "type": "array", "items": { "$ref": "#/components/schemas/Customer" } },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer" },
                        "page": { "type": "integer" },
                        "pages": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "summary": "Create customer",
        "description": "Create a new customer",
        "tags": ["Customers"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["first_name", "last_name"],
                "properties": {
                  "first_name": { "type": "string" },
                  "last_name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "phone": { "type": "string" },
                  "address": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Customer created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "customer": { "$ref": "#/components/schemas/Customer" }
                  }
                }
              }
            }
          },
          "409": { "$ref": "#/components/responses/ConflictError" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/customers/{id}": {
      "get": {
        "summary": "Get customer by ID",
        "description": "Retrieve a specific customer",
        "tags": ["Customers"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Customer retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "customer": { "$ref": "#/components/schemas/Customer" }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "put": {
        "summary": "Update customer",
        "description": "Update customer information",
        "tags": ["Customers"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "first_name": { "type": "string" },
                  "last_name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "phone": { "type": "string" },
                  "address": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "customer": { "$ref": "#/components/schemas/Customer" }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "409": { "$ref": "#/components/responses/ConflictError" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "delete": {
        "summary": "Delete customer",
        "description": "Delete a customer",
        "tags": ["Customers"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Customer deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/api/locations": {
      "get": {
        "summary": "Get all locations",
        "description": "Retrieve all locations for the tenant",
        "tags": ["Locations"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Locations retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "locations": { "type": "array", "items": { "$ref": "#/components/schemas/Location" } }
                  }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "summary": "Create location",
        "description": "Create a new location (admin/manager only)",
        "tags": ["Locations"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "address"],
                "properties": {
                  "name": { "type": "string" },
                  "address": { "type": "string" },
                  "phone": { "type": "string" },
                  "tax_rate": { "type": "number", "format": "decimal", "default": 0 },
                  "pricing_multiplier": { "type": "number", "format": "decimal", "default": 1 },
                  "operating_hours": { "type": "object" },
                  "timezone": { "type": "string", "default": "UTC" },
                  "currency": { "type": "string", "default": "USD" },
                  "is_active": { "type": "boolean", "default": true }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Location created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "location": { "$ref": "#/components/schemas/Location" }
                  }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    }
  },
  "components": {
    "responses": {
...
      "NotFoundError": {
        "description": "Not Found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "ConflictError": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "InternalError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    }
  }
}